Create movie rating filter component
