+++ cinema_frontend/src/components/Filters/Filters.jsx
@@ -5,21 +5,14 @@ import {
     PopoverContent,
 } from "@/components/ui/popover";
 import { Button } from "@/components/ui/button";
-import { Slider } from "@/components/ui/slider";
-import { Select, SelectTrigger, SelectContent, SelectItem } from "@/components/ui/select";
-import { Label } from "@/components/ui/label";
-import { Input } from "@/components/ui/input";
 import { ChevronDown } from "lucide-react";
 import GenreFilter from "./GenreFilter/GenreFilter";
 import YearFilter from "./YearFilter/YearFilter";
 import DurationFilter from "./DurationFilter/DurationFilter";
 import AgeRatingFilter from "./AgeRatingFilter/AgeRatingFilter";
-
-const ratingsRange = [0, 10]; 
+import MovieRatingFilter from "./MovieRatingFilter/MovieRatingFilter";
 
 const Filters = () => {
-    const [movieRating, setMovieRating] = useState([0, 10]); 
-
     return (
         <Popover>
             <PopoverTrigger>
@@ -42,22 +35,7 @@ const Filters = () => {
                     <AgeRatingFilter />
 
                     {/* Movie Rating Filter */}
-                    <div>
-                        <Label htmlFor="movieRating" className="font-semibold">
-                            Movie Rating (0 to 10)
-                        </Label>
-                        <Slider
-                            value={movieRating}
-                            min={ratingsRange[0]}
-                            max={ratingsRange[1]}
-                            step={0.1}
-                            onChange={setMovieRating}
-                        />
-                        <div className="flex justify-between text-sm mt-2">
-                            <span>{movieRating[0]}</span>
-                            <span>{movieRating[1]}</span>
-                        </div>
-                    </div>
+                    <MovieRatingFilter />
 
                     {/* Apply Filters Button */}
                     <div className="text-right">
